General Refactor and beeAPI implemented
