+++ BeeAppy/src/components/BeeButton.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import { TouchableHighlight, View, Text } from 'react-native';
-import BeeAPI from '../BeeAPI';
+import BeeAPI from './BeeAPI';
 
-const api = new BeeAPI()
+const api = new BeeAPI();
 
 export default class BeeButton extends React.Component {
+
   render() {
     return(
       <View>
         <TouchableHighlight
-          onPress = {BeeAPI.POSTbee}>
+          onPress = {api.POSTbee}>
           <Text>BeeButton</Text>
         </TouchableHighlight>
       </View>
